refactor: cookie util 정리
