+++ src/utils/cookie/serverUtils.ts
@@ -1,0 +1,5 @@
+import { cookies } from 'next/headers';
+
+export const getServerCookie = (name: string) => {
+  return cookies().get(name)?.value;
+};
